// src/types/systems/colors/color-obj-controller.d.ts
import { Color } from '@material-ui/core';

export interface ColorObjects {
            [key: string]: Color;
            red: Color;
            pink: Color;
            purple: Color;
            deepPurple: Color;
            indigo: Color;
            blue: Color;
            lightBlue: Color;
            cyan: Color;
            teal: Color;
            green: Color;
            lightGreen: Color;
            lime: Color;
            yellow: Color;
            amber: Color;
            orange: Color;
            deepOrange: Color;
            brown: Color;
            grey: Color;
            blueGrey: Color;
}